append project local archives to 'archives' file
The content of the archives mentioned in a pkg's 'archives' file is
subject to shadowing. Sculpt lets subsequent entries override previous
ones. `goa run` does the same and, moreover, processes the project local
src and raw archives at the very last. However, when publishing a pkg
archive, Goa placed the project's src and raw archives at the very
beginning of the 'archives' file. This commit corrects this.

genodelabs/goa#39