add --depot-retain switch
This serves as an alternative to the --depot-overwrite switch and
instructs Goa to keep existing depot content but continue command
execution. This is useful in two scenarios:

a) When republishing an index project, whose dependent projects have
   already been exported.
b) When exporting a project that has already been published.

Fixes genodelabs/goa#71