util: add `query_raw_from_file`
Changed `query_from_file` to execute xmllint with `--format`. This
removes newlines from the XML file so that each found XML node is
returned as a single line. This becomes important if the queried nodes
contain subnodes. Without `--format` the subnodes may appear on a
separate line and mess up the returned string of `query_from_file`.

The former behaviour is now implemented by `query_raw_from_file`, which is only
used for acquiring the config XML.